move helper which were used only for action tests
